* fix segfault in OrthoRenderCGO   https://sourceforge.net/p/pymol/bugs/137/ * importing.load_coords * improve util.chainbow * fix plugin manager PyMOLWiki URL install: strip whitespace

+++ trunk/pymol/layer1/PConv.c
@@ -482,7 +482,7 @@ int PConvPyListToFloatArray(PyObject * obj, float **f)
     *f = NULL;
     ok = false;
   } else {
-    l = PyList_Size(obj);
+    l = (int) PyList_Size(obj);
     if(!l)
       ok = -1;
     else
